ccsd works
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@247 6a214465-e6f0-41bc-91e6-de896bb96b5e

+++ src/lib/libqt/qt.h
@@ -112,7 +112,7 @@ void C_DGEMM(char transa, char transb, int m, int n, int k,
              double beta, double *C, int ncc);
 void C_DROT(int ntot, double *x, int incx, double *y, int incy,
              double costheta, double sintheta);
-void C_DSCAL(long int len, double alpha, double *vec, int inc);
+void C_DSCAL(int len, double alpha, double *vec, int inc);
 void C_DGEMV(char transa, int m, int n, double alpha, double *A,
              int nca, double *X, int inc_x, double beta, double *Y,
              int inc_y);
